refactor: removed duplicated examples
