+++ .gitignore
@@ -1,3 +1,4 @@
+./dist/**/*
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
 ##
